+++ RELEASE_NOTES.md
@@ -8,7 +8,7 @@
 
 - Icons migration step 1 (NON BREAKING ... for now and until next major release. But we're moving away from Google's material-icons)
   - Import Atlas icons (via CDN via CSS) for Atlas icons
-  - Add custom icons "Swepay-icons": a few svg icons saved locally to complement where missing from Atlas icons set (called Swepay-icons but supported for both Swedbankpay AND PayEx sides)
+  - Add custom icons "Swepay-icons": svg icons saved locally to complement the ones missing from Atlas icons set (called Swepay-icons but supported for both Swedbankpay AND PayEx sides)
   - Add support in the components styles for the Atlas icons and our new custom made "Swepay-icons"
 
 ## Maintenance